requirements.txt
asgiref==3.2.3
Django==3.0.2
django-redis==4.11.0
djangorestframework==3.11.0
elasticsearch==7.5.1
mongoengine==0.19.1
mysqlclient==1.4.6
pika==1.1.0
pymongo==3.10.1
pytz==2019.3
redis==3.4.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
